feat: mobile view, update styles, file breakdown
